refactor: extract sandbox TOS config helpers
(cherry picked from commit bbd49245e6db74cc65864ef5c4aeb343ced6fb0b)
